Moved TPSConnectorCLI classes
The TPSConnectorCLI classes have been moved into
com.netscape.cmstools.tks since they are used to
manage the TPS connector in TKS.
