+++ base/java-tools/src/com/netscape/cmstools/tks/TPSConnectorModCLI.java
@@ -1,4 +1,4 @@
-package com.netscape.cmstools.system;
+package com.netscape.cmstools.tks;
 
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.Option;
